+++ icat/dumpfile.py
@@ -94,7 +94,7 @@ class DumpFileReader():
         file name.
 
     .. versionchanged:: 1.0.0
-        The `infile` parameter also accepts a :class:`~pathlib.Path`
+        the `infile` parameter also accepts a :class:`~pathlib.Path`
         object.
     """
 
@@ -194,7 +194,7 @@ class DumpFileWriter():
         :class:`~pathlib.Path` or a :class:`str` with a file name.
 
     .. versionchanged:: 1.0.0
-        The `outfile` parameter also accepts a :class:`~pathlib.Path`
+        the `outfile` parameter also accepts a :class:`~pathlib.Path`
         object.
     """