+++ CHANGES.rst
@@ -16,6 +16,8 @@ New features
 Bug fixes and minor changes
 ---------------------------
 
++ `#126`_, `#127`_: Update outdated documentation.
+
 + `#112`_, `#118`_: Extend icatdata XSD adding extra attributes to
   reference objects.
 
@@ -44,6 +46,8 @@ Bug fixes and minor changes
 .. _#121: https://github.com/icatproject/python-icat/pull/121
 .. _#123: https://github.com/icatproject/python-icat/pull/123
 .. _#124: https://github.com/icatproject/python-icat/pull/124
+.. _#126: https://github.com/icatproject/python-icat/issues/126
+.. _#127: https://github.com/icatproject/python-icat/pull/127
 .. _#128: https://github.com/icatproject/python-icat/pull/128